# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trading Buddy API",
    "description": "A股量化交易数据基础设施",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/dashboard/gainers": {
      "get": {
        "tags": [
          "看板"
        ],
        "summary": "Get Top Gainers",
        "description": "获取涨幅榜",
        "operationId": "get_top_gainers_api_dashboard_gainers_get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "交易日期；省略则用库中最新交易日",
              "title": "Trade Date"
            },
            "description": "交易日期；省略则用库中最新交易日"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "description": "返回数量",
              "default": 10,
              "title": "Limit"
            },
            "description": "返回数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Top Gainers Api Dashboard Gainers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/losers": {
      "get": {
        "tags": [
          "看板"
        ],
        "summary": "Get Top Losers",
        "description": "获取跌幅榜",
        "operationId": "get_top_losers_api_dashboard_losers_get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "交易日期；省略则用库中最新交易日",
              "title": "Trade Date"
            },
            "description": "交易日期；省略则用库中最新交易日"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "description": "返回数量",
              "default": 10,
              "title": "Limit"
            },
            "description": "返回数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Top Losers Api Dashboard Losers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/turnover": {
      "get": {
        "tags": [
          "看板"
        ],
        "summary": "Get High Turnover",
        "description": "获取成交额排行榜：全市场（有 ``stock_info`` 且交易中）在指定日按 ``amount`` 降序 Top ``limit``。",
        "operationId": "get_high_turnover_api_dashboard_turnover_get",
        "parameters": [
          {
            "name": "trade_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "交易日期；省略则用库中最新交易日（与涨跌榜口径一致）",
              "title": "Trade Date"
            },
            "description": "交易日期；省略则用库中最新交易日（与涨跌榜口径一致）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get High Turnover Api Dashboard Turnover Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}